Support For New User Sign Up
This commit implements the ability to sign up as a new user, and also
includes some safety checks. The checks occur on the php side to ensure
username and password are populated, not a duplicate of what is already
in the system, and a confirmation that the user has been created.

Additionally, the php file is used in sync with both the HTML site and
Android application.